LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.font.parse.invalid.worker-expected.txt


FAIL 2d assert_equals: ctx.font === '20px serif' (got bogus[string], expected 20px serif[string]) expected "20px serif" but got "bogus"
